+++ sdk/networkfunction/azure-mgmt-networkfunction/CHANGELOG.md
@@ -1,5 +1,5 @@
 # Release History
 
-## 1.0.0b1 (2022-11-17)
+## 1.0.0b1 (2022-11-18)
 
 * Initial Release
