+++ sdk/networkfunction/azure-mgmt-networkfunction/_meta.json
@@ -1,5 +1,5 @@
 {
-  "commit": "b32e1896f30e6ea155449cb49719a6286e32b961",
+  "commit": "4f8a52119114fdc8159fcc556ec580b090241f33",
   "repository_url": "https://github.com/Azure/azure-rest-api-specs",
   "autorest": "3.9.2",
   "use": [
