+++ sdk/networkfunction/azure-mgmt-networkfunction/azure/mgmt/networkfunction/__init__.py
@@ -6,7 +6,7 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-from ._azure_traffic_collector import AzureTrafficCollector
+from ._traffic_collector_mgmt_client import TrafficCollectorMgmtClient
 from ._version import VERSION
 
 __version__ = VERSION
@@ -19,7 +19,7 @@ except ImportError:
 from ._patch import patch_sdk as _patch_sdk
 
 __all__ = [
-    "AzureTrafficCollector",
+    "TrafficCollectorMgmtClient",
 ]
 __all__.extend([p for p in _patch_all if p not in __all__])
