+++ sdk/networkfunction/azure-mgmt-networkfunction/azure/mgmt/networkfunction/aio/operations/_network_function_operations.py
@@ -44,7 +44,7 @@ class NetworkFunctionOperations:
         **DO NOT** instantiate this class directly.
 
         Instead, you should access the following operations through
-        :class:`~azure.mgmt.networkfunction.aio.AzureTrafficCollector`'s
+        :class:`~azure.mgmt.networkfunction.aio.TrafficCollectorMgmtClient`'s
         :attr:`network_function` attribute.
     """
