+++ sdk/networkfunction/azure-mgmt-networkfunction/azure/mgmt/networkfunction/models/_traffic_collector_mgmt_client_enums.py
