+++ sdk/networkfunction/azure-mgmt-networkfunction/azure/mgmt/networkfunction/operations/_azure_traffic_collectors_by_resource_group_operations.py
@@ -75,7 +75,7 @@ class AzureTrafficCollectorsByResourceGroupOperations:
         **DO NOT** instantiate this class directly.
 
         Instead, you should access the following operations through
-        :class:`~azure.mgmt.networkfunction.AzureTrafficCollector`'s
+        :class:`~azure.mgmt.networkfunction.TrafficCollectorMgmtClient`'s
         :attr:`azure_traffic_collectors_by_resource_group` attribute.
     """
