+++ sdk/networkfunction/azure-mgmt-networkfunction/azure/mgmt/networkfunction/operations/_azure_traffic_collectors_by_subscription_operations.py
@@ -73,7 +73,7 @@ class AzureTrafficCollectorsBySubscriptionOperations:
         **DO NOT** instantiate this class directly.
 
         Instead, you should access the following operations through
-        :class:`~azure.mgmt.networkfunction.AzureTrafficCollector`'s
+        :class:`~azure.mgmt.networkfunction.TrafficCollectorMgmtClient`'s
         :attr:`azure_traffic_collectors_by_subscription` attribute.
     """
