+++ sdk/networkfunction/azure-mgmt-networkfunction/azure/mgmt/networkfunction/operations/_azure_traffic_collectors_operations.py
@@ -185,7 +185,7 @@ class AzureTrafficCollectorsOperations:
         **DO NOT** instantiate this class directly.
 
         Instead, you should access the following operations through
-        :class:`~azure.mgmt.networkfunction.AzureTrafficCollector`'s
+        :class:`~azure.mgmt.networkfunction.TrafficCollectorMgmtClient`'s
         :attr:`azure_traffic_collectors` attribute.
     """
