+++ sdk/networkfunction/azure-mgmt-networkfunction/azure/mgmt/networkfunction/operations/_collector_policies_operations.py
@@ -240,7 +240,7 @@ class CollectorPoliciesOperations:
         **DO NOT** instantiate this class directly.
 
         Instead, you should access the following operations through
-        :class:`~azure.mgmt.networkfunction.AzureTrafficCollector`'s
+        :class:`~azure.mgmt.networkfunction.TrafficCollectorMgmtClient`'s
         :attr:`collector_policies` attribute.
     """
