+++ sdk/networkfunction/azure-mgmt-networkfunction/azure/mgmt/networkfunction/operations/_network_function_operations.py
@@ -66,7 +66,7 @@ class NetworkFunctionOperations:
         **DO NOT** instantiate this class directly.
 
         Instead, you should access the following operations through
-        :class:`~azure.mgmt.networkfunction.AzureTrafficCollector`'s
+        :class:`~azure.mgmt.networkfunction.TrafficCollectorMgmtClient`'s
         :attr:`network_function` attribute.
     """
